refactor: move KafkaThread to internal
