+++ clients/src/main/java/org/apache/kafka/common/utils/Shell.java
@@ -16,6 +16,7 @@
  */
 package org.apache.kafka.common.utils;
 
+import org.apache.kafka.common.utils.internals.KafkaThread;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
