+++ server-common/src/main/java/org/apache/kafka/queue/KafkaEventQueue.java
@@ -18,7 +18,7 @@
 package org.apache.kafka.queue;
 
 import org.apache.kafka.common.errors.TimeoutException;
-import org.apache.kafka.common.utils.KafkaThread;
+import org.apache.kafka.common.utils.internals.KafkaThread;
 import org.apache.kafka.common.utils.LogContext;
 import org.apache.kafka.common.utils.Time;
