+++ server-common/src/main/java/org/apache/kafka/server/util/KafkaScheduler.java
@@ -16,7 +16,7 @@
  */
 package org.apache.kafka.server.util;
 
-import org.apache.kafka.common.utils.KafkaThread;
+import org.apache.kafka.common.utils.internals.KafkaThread;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
