+++ server-common/src/main/java/org/apache/kafka/server/util/timer/SystemTimer.java
@@ -16,7 +16,7 @@
  */
 package org.apache.kafka.server.util.timer;
 
-import org.apache.kafka.common.utils.KafkaThread;
+import org.apache.kafka.common.utils.internals.KafkaThread;
 import org.apache.kafka.common.utils.ThreadUtils;
 import org.apache.kafka.common.utils.Time;
